# contracts/fastapi-openapi-v0.1.0.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tender Knowledge Hub",
    "version": "0.1.0"
  },
  "paths": {
...
    "/dashboard/vault/notes/{note_path}": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Vault Note",
        "operationId": "vault_note_dashboard_vault_notes__note_path__get",
        "parameters": [
          {
            "name": "note_path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Note Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Vault Note Dashboard Vault Notes  Note Path  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/files/{document_id}": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Download Document",
        "operationId": "download_document_dashboard_files__document_id__get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/files/{document_id}/view": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "View Document",
        "operationId": "view_document_dashboard_files__document_id__view_get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/tree-files/{root_key}/{file_path}": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Download Tree File",
        "operationId": "download_tree_file_dashboard_tree_files__root_key___file_path__get",
        "parameters": [
          {
            "name": "root_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Root Key"
            }
          },
          {
            "name": "file_path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "File Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/tree-file-view/{root_key}/{file_path}": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "View Tree File",
        "operationId": "view_tree_file_dashboard_tree_file_view__root_key___file_path__get",
        "parameters": [
          {
            "name": "root_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Root Key"
            }
          },
          {
            "name": "file_path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "File Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/upload": {
      "post": {
        "tags": [
          "dashboard"
        ],
        "summary": "Upload Document",
        "operationId": "upload_document_dashboard_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_dashboard_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_upload_document_dashboard_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "internal_unit": {
            "type": "string",
            "title": "Internal Unit"
          },
          "organization": {
            "type": "string",
            "title": "Organization"
          },
          "tender_id": {
            "type": "string",
            "title": "Tender Id",
            "default": ""
          },
          "year": {
            "type": "integer",
            "title": "Year"
          },
          "caption": {
            "type": "string",
            "title": "Caption",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "file",
          "internal_unit",
          "organization",
          "year"
        ],
        "title": "Body_upload_document_dashboard_upload_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}